Feature/email send

+++ components/cart/List.vue
@@ -1,14 +1,14 @@
 <script setup lang="ts">
+import { useToast } from 'vue-toastification'
 import { ArrowLongLeftIcon, XMarkIcon } from '@heroicons/vue/24/outline'
 import type { Cart, Services } from '~/types/general'
-import { usePayment } from '~/composables/usePayment'
+import { formatPrice } from '~/utils/price'
 
+const toast = useToast()
 const { user: authUser } = useAuth()
 const cartData = ref<Cart[]>([])
 const servicesData = ref<Services[]>([])
 const loading = ref(false)
-const showToast = ref(false)
-const toastMessage = ref('')
 const { appointmentForm, isAlreadyExistedModal } = usePayment()
 const selectedTime = ref<string>('')
 const consultationType = ref<string>('office')
@@ -106,24 +106,19 @@ async function proceedToPayment() {
 }
 
 async function deleteService(service: Services) {
-  showToast.value = true
-  toastMessage.value = 'Removed Service Successfully!'
+  if (loading.value) return
+  
+  loading.value = true
+
+  toast.success('Cart Removed Successfully!')
 
   const { error } = await $useCustomFetch(`/api/services/${service.id}`, { 
     method: 'DELETE',
   })
 
-  showToast.value = false
+  loading.value = false
 
   await fetchServices()
-}
-
-function formatPrice(price: number) {
-  return new Intl.NumberFormat('en-PH', {
-    style: 'currency',
-    currency: 'PHP',
-    minimumFractionDigits: 2,
-  }).format(price)
 }
 
 const isButtonDisabled = computed<boolean>(() => {
@@ -163,6 +158,12 @@ await fetchCart()
           </div>
 
           <XMarkIcon class="w-6 h-6 stroke-custom-brown-500 cursor-pointer" @click="deleteService(service)" />
+        </div>
+
+        <div v-if="!servicesData.length" class="flex flex-col items-center w-full space-y-4">
+          <img src="/images/empty-cart.svg" />
+
+          <p claas="text-sm text-custom-brown-500 font-medium">Your cart is empty</p>
         </div>
       </div>
 
@@ -218,6 +219,4 @@ await fetchCart()
   </div>
 
   <CartAlreadyExistedModal />
-  
-  <BaseToast :show="showToast" :message="toastMessage" />
 </template>
